Split client profile source context mapping
